chatbot-api/requirements.txt
matplotlib
numpy
scikit-learn
pydantic
google-generativeai
langextract
python-dotenv
Flask
requests
folium
